+++ lib/conform/schema.ex
@@ -164,7 +164,7 @@ defmodule Conform.Schema do
           case valid_import?(i) do
             true  -> i
             false ->
-              Conform.Utils.warn "Schema imports #{i}, but #{i} could not be loaded."
+              Conform.Logger.warn "Schema imports #{i}, but #{i} could not be loaded."
               nil
           end
         end) |> Enum.filter(fn nil -> false; _ -> true end)
@@ -204,13 +204,13 @@ defmodule Conform.Schema do
           e when is_atom(e) ->
             case get_extends_schema(e, path) do
               nil ->
-                Conform.Utils.warn "Schema extends #{e}, but the schema for #{e} was not found."
+                Conform.Logger.warn "Schema extends #{e}, but the schema for #{e} was not found."
                 nil
               {schema_path, contents} ->
                 contents |> parse! |> from(schema_path)
             end
           e ->
-            Conform.Utils.warn "Invalid extends value: #{e}. Only application names as atoms are permitted."
+            Conform.Logger.warn "Invalid extends value: #{e}. Only application names as atoms are permitted."
             nil
         end) |> Enum.filter(fn nil -> false; _ -> true end)
         # Merge them onto the base schema in the order provided
@@ -241,13 +241,12 @@ defmodule Conform.Schema do
   @spec coalesce() :: schema
   def coalesce do
     # Get schemas from all dependencies
-    proj_config = [build_path: Mix.Project.build_path, umbrella?: Mix.Project.umbrella?]
     # Merge schemas for all deps
     Mix.Dep.loaded([])
     |> Enum.map(fn %Mix.Dep{app: app, opts: opts} ->
-         Mix.Project.in_project(app, opts[:dest], proj_config, fn _ -> load!(app) end)
+         Mix.Project.in_project(app, opts[:dest], opts, fn _ -> load!(app) end)
        end)
-    |> coalesce
+    |> coalesce()
   end
 
   @doc """
@@ -365,7 +364,7 @@ defmodule Conform.Schema do
   def empty, do: %Schema{}
 
   defp to_schema([]),     do: %Schema{}
-  defp to_schema(config), do: to_schema(config, %Schema{})
+  defp to_schema(config), do: to_schema(Macro.escape(config), %Schema{})
   defp to_schema([], schema), do: schema
   defp to_schema([{app, settings} | config], schema) do
     mappings = settings