+++ lib/mix/tasks/conform.effective.ex
@@ -29,8 +29,7 @@ defmodule Mix.Tasks.Conform.Effective do
   """
   @shortdoc "Print the effective configuration for the current project"
 
-  use    Mix.Task
-  import Conform.Utils
+  use Mix.Task
 
   # Perform some actions within the context of a specific mix environment
   defp with_env(env, fun) do
@@ -49,16 +48,18 @@ defmodule Mix.Tasks.Conform.Effective do
     args = parse_args(argv)
     cond do
       Mix.Project.umbrella? && args.options.type == :app ->
-        config = [build_path: Mix.Project.build_path]
         app = args.options.app
-        for %Mix.Dep{app: ^app, opts: opts} <- Mix.Dep.Umbrella.loaded do
-          Mix.Project.in_project(app, opts[:path], config, fn _ -> do_run(args) end)
+        dep = Enum.find(Mix.Dep.Umbrella.loaded, fn %Mix.Dep{app: ^app} -> true; _ -> false end)
+        case dep do
+          %Mix.Dep{opts: opts} ->
+            Mix.Project.in_project(app, opts[:path], opts, fn _ -> do_run(args) end)
+          _ ->
+            Conform.Logger.error "#{app} could not be found"
         end
       Mix.Project.umbrella? ->
-        config = [build_path: Mix.Project.build_path]
         for %Mix.Dep{app: app, opts: opts} <- Mix.Dep.Umbrella.loaded do
-          IO.puts "in #{app}.."
-          Mix.Project.in_project(app, opts[:path], config, fn _ -> do_run(args) end)
+          Conform.Logger.info "in #{app}"
+          Mix.Project.in_project(app, opts[:path], opts, fn _ -> do_run(args) end)
         end
       :else ->
         do_run(args)
@@ -67,62 +68,86 @@ defmodule Mix.Tasks.Conform.Effective do
 
   defp do_run(args) do
     Mix.Tasks.Loadpaths.run([])
-    app       = Mix.Project.config |> Keyword.get(:app)
+
+    project_config = Mix.Project.config()
+    app = project_config[:app]
     config_path = Path.join(File.cwd!, "config")
-    conf_path = case File.exists?(Path.join(config_path, "#{app}.#{args.env}.conf")) do
-                  true  -> Path.join(config_path, "#{app}.#{args.env}.conf")
-                  false -> Path.join(config_path, "#{app}.conf")
-                end
+    conf_path =
+      case File.exists?(Path.join(config_path, "#{app}.#{args.env}.conf")) do
+        true  -> Path.join(config_path, "#{app}.#{args.env}.conf")
+        false -> Path.join(config_path, "#{app}.conf")
+      end
     # Load the base configuration from config.exs if it exists, and validate it
     # If config.exs doesn't exist, proceed if a .conf file exists, otherwise there
     # is no configuration to display
-    config = case File.exists?("config/config.exs") do
-      true ->
+    config =
+      if File.exists?("config/config.exs") do
         # Switch environments when reading the config
         with_env args.env, fn ->
           Path.join([File.cwd!, "config", "config.exs"]) |> Mix.Config.read!
         end
-      false -> []
-    end
+      else
+        []
+      end
     # Read .conf
-    conf = case File.exists?(conf_path) do
-      false ->
-        error "You must have a .conf file in order to run conform.effective!"
-        exit(:normal)
-      true  ->
+    conf =
+      if File.exists?(conf_path) do
         case Conform.Conf.from_file(conf_path) do
+          {:error, reason} when is_binary(reason) ->
+            Conform.Logger.error "Failed to parse .conf!\nError: #{reason}"
           {:error, reason} ->
-            error reason
-            exit(:normal)
+            Conform.Logger.error "Failed to parse .conf!\nError: #{inspect reason}"
           {:ok, table} ->
             table
         end
-    end
+      else
+        Conform.Logger.debug "No .conf file found - defaults from schema and Mix config will be used"
+          {:ok, table} = Conform.Conf.from([])
+          table
+      end
     # Load merged schemas
-    schema = Conform.Schema.schema_path(app) |> Conform.Schema.load!
+    schema =
+      app
+      |> Conform.Schema.schema_path
+      |> Conform.Schema.load!
     # Translate .conf -> config, using settings from config if one is
     # not provided in the .conf. If no setting is present in either
     # the config, or the .conf, the default from the schema is used.
     effective = Conform.Translate.to_config(schema, config, conf)
     # Print the configuration as requested
-    output = case args.options.type do
-      :schema -> Conform.Schema.stringify(schema)
-      :all    -> Conform.SysConfig.prettify(effective)
-      :app    ->
+    output_type = args.options.output
+    case args.options.type do
+      :schema when output_type == :stdout ->
+        schema |> Conform.Schema.stringify |> IO.puts
+      :schema when is_binary(output_type) ->
+        output = schema |> Conform.Schema.stringify
+        File.write!(output_type, output)
+      :all when output_type == :stdout ->
+        Conform.SysConfig.print(effective)
+      :all when is_binary(output_type) ->
+        output = Conform.SysConfig.prettify(effective)
+        File.write!(output_type, output)
+      :app ->
         case Keyword.get(effective, args.options.app) do
           nil ->
-            notice "App not found!"
-            ""
-          app_config ->
-            Conform.SysConfig.prettify(app_config)
+            Conform.Logger.warn "App not found!"
+            if is_binary(output_type) do
+              File.touch(output_type)
+            end
+          app_config when output_type == :stdout ->
+            Conform.SysConfig.print(app_config)
+          app_config when is_binary(output_type) ->
+            output = Conform.SysConfig.prettify(app_config)
+            File.write!(output_type, output)
         end
-      :key ->
-        effective |> extract_key(args.options.key) |> Conform.SysConfig.prettify
-    end
-
-    case args.options.output do
-      :stdout -> IO.puts(output)
-      path    -> File.write!(path, output)
+      :key when output_type == :stdout ->
+        config = effective |> extract_key(args.options.key)
+        Conform.SysConfig.print(config)
+      :key when is_binary(output_type) ->
+        config = effective |> extract_key(args.options.key) |> Conform.SysConfig.prettify
+        File.write!(output_type, config)
+      _ ->
+        Conform.Logger.error "Invalid output: #{inspect output_type}"
     end
   end
 
@@ -134,8 +159,7 @@ defmodule Mix.Tasks.Conform.Effective do
       config |> get_in(spec)
     rescue
       Protocol.UndefinedError ->
-        notice "The value for some key in #{path} doesn't implement the Access protocol!"
-        exit(:normal)
+        Conform.Logger.error "The value for some key in #{path} doesn't implement the Access protocol!"
     end
   end
 
@@ -149,8 +173,7 @@ defmodule Mix.Tasks.Conform.Effective do
 
     mutually_exclusive_opts = Enum.filter([app: app, key: key, schema: schema], fn({_, val}) -> val != nil end)
     if Enum.count(mutually_exclusive_opts) > 1 do
-      error "the options #{mutually_exclusive_opts} are mutually exclusive"
-      exit(:normal)
+      Conform.Logger.error "the options #{mutually_exclusive_opts} are mutually exclusive"
     end
 
     # populate options