+++ lib/mix/tasks/conform.new.ex
@@ -9,16 +9,14 @@ defmodule Mix.Tasks.Conform.New do
   """
   @shortdoc "Create a new .schema.exs file for configuring your app with conform"
 
-  use    Mix.Task
-  import Conform.Utils
+  use Mix.Task
 
   def run(_args) do
     Mix.Tasks.Loadpaths.run([])
     if Mix.Project.umbrella? do
-      config = [build_path: Mix.Project.build_path]
       # Execute task for each project in the umbrella
       for %Mix.Dep{app: app, opts: opts} <- Mix.Dep.Umbrella.loaded do
-        Mix.Project.in_project(app, opts[:path], config, fn _ ->
+        Mix.Project.in_project(app, opts[:path], opts, fn _ ->
           do_run(app, Path.expand("config/config.exs"))
         end)
       end
@@ -32,24 +30,26 @@ defmodule Mix.Tasks.Conform.New do
     # Load the configuration for this app, and convert configuration to schema format
     output_path = Conform.Schema.schema_path(app)
     # Make sure we want to proceed if a schema already exists
-    continue? = case File.exists?(output_path) do
-      true  -> confirm_overwrite?(output_path)
-      false -> true
-    end
+    continue? =
+      if File.exists?(output_path) do
+        confirm_overwrite?(output_path)
+      else
+        true
+      end
     if continue? do
       # Ensure output directory exists
       output_path |> Path.dirname |> File.mkdir_p!
       if File.exists?(config_path) do
         # Load existing config and convert it to quoted schema terms
-        config = Mix.Config.read!(config_path) |> Macro.escape
+        config = Mix.Config.read!(config_path)
         schema = Conform.Schema.from_config(config)
         # Write the generated schema to `output_path`
         Conform.Schema.write_quoted(schema, output_path)
-        info "The schema for your project has been placed in #{Path.relative_to_cwd(output_path)}"
+        Conform.Logger.success "The schema for your project has been placed in #{Path.relative_to_cwd(output_path)}"
       else
-        warn "Your project does not currently have any configuration!"
+        Conform.Logger.warn "Your project does not currently have any configuration!"
         Conform.Schema.write_quoted(Conform.Schema.empty, output_path)
-        info "An empty schema has been placed in #{Path.relative_to_cwd(output_path)}"
+        Conform.Logger.success "An empty schema has been placed in #{Path.relative_to_cwd(output_path)}"
       end
     end
   end
@@ -63,5 +63,4 @@ defmodule Mix.Tasks.Conform.New do
     IO.puts IO.ANSI.reset
     confirmed?
   end
-
 end