+++ test/config_test.exs
@@ -23,16 +23,17 @@ defmodule ConfigTest do
          complex_list: [first: [age: 20, username: "username1"],
                         second: [age: 40, username: "username2"]],
          db: [hosts: [{"127.0.0.1", "8000"}, {"127.0.0.2", "8001"}]],
+         max_demand: 40,
          nodelist: [:'a@bar', :'b@bar'],
          some_val: :foo
        ]]
-    assert expected == sysconfig
+    assert expected == Conform.Utils.sort_kwlist(sysconfig)
   end
 
   test "issue #85" do
     path = Path.join(["test", "configs", "issue_85.exs"])
     output_path = Path.join(["test", "configs", "issue_85.schema.exs"])
-    config_raw = path |> Mix.Config.read! |> Macro.escape
+    config_raw = path |> Mix.Config.read!
     config = path |> Mix.Config.read!
     assert [rocket: _] = config
 
@@ -45,7 +46,7 @@ defmodule ConfigTest do
   test "issue #114" do
     path = Path.join(["test", "configs", "issue_114.exs"])
     output_path = Path.join(["test", "configs", "issue_114.schema.exs"])
-    config_raw = path |> Mix.Config.read! |> Macro.escape
+    config_raw = path |> Mix.Config.read!
 
     schema = Conform.Schema.from_config(config_raw)
     assert %Schema{} = schema
@@ -57,7 +58,7 @@ defmodule ConfigTest do
   test "issue #75" do
     path = Path.join(["test", "configs", "raw_binary.exs"])
     output_path = Path.join(["test", "configs", "raw_binary.schema.exs"])
-    config_raw = path |> Mix.Config.read! |> Macro.escape
+    config_raw = path |> Mix.Config.read!
     config = path |> Mix.Config.read!
 
     assert [my_app: _] = config
@@ -79,7 +80,7 @@ defmodule ConfigTest do
 
   test "logger example" do
     path = Path.join(["test", "configs", "logger.exs"])
-    config_raw = path |> Mix.Config.read! |> Macro.escape
+    config_raw = path |> Mix.Config.read!
     config = path |> Mix.Config.read!
     assert [logger: [backends: [:console, {ExSyslog, :exsyslog_error}, {ExSyslog, :exsyslog_debug}]]] = config
     schema = Conform.Schema.from_config(config_raw)
@@ -107,7 +108,7 @@ defmodule ConfigTest do
 
   test "can translate config.exs containing nested lists to schema" do
     path   = Path.join(["test", "configs", "nested_list.exs"])
-    config = path |> Mix.Config.read! |> Macro.escape
+    config = path |> Mix.Config.read!
     schema = Conform.Schema.from_config(config)
     assert %Schema{extends: [], import: [],
             mappings: [%Mapping{
@@ -129,7 +130,7 @@ defmodule ConfigTest do
 
   test "can translate config.exs containing a single nested list to schema" do
     path   = Path.join(["test", "configs", "single_nested_list.exs"])
-    config = path |> Mix.Config.read! |> Macro.escape
+    config = path |> Mix.Config.read!
     schema = Conform.Schema.from_config(config)
     assert %Schema{extends: [], import: [],
                    mappings: [%Mapping{
@@ -151,12 +152,23 @@ defmodule ConfigTest do
 
   test "can translate config.exs + schema + conf with nested lists to sys.config" do
     path   = Path.join(["test", "configs", "nested_list.exs"])
-    config_raw = path |> Mix.Config.read! |> Macro.escape
+    config_raw = path |> Mix.Config.read!
     config = path |> Mix.Config.read!
     schema = Conform.Schema.from_config(config_raw)
     {:ok, conf} = Path.join(["test", "confs", "nested_list.conf"]) |> Conform.Conf.from_file
     sysconfig = Conform.Translate.to_config(schema, config, conf)
     assert [my_app: [rx_pattern: [~r/[A-Z]+/],
                      sublist: [[opt1: "val1", opt2: "val two"], [opt1: "val3", opt2: "val-4"]]]] == sysconfig
+  end
+
+  test "raises runtime error if env var value is missing and there is no default value" do
+    schema_path = Path.join(["test", "schemas", "env_var.schema.exs"])
+    schema = Conform.Schema.load!(schema_path)
+    conf_path = Path.join(["test", "confs", "test.conf"])
+    {:ok, conf} = Conform.Conf.from_file(conf_path)
+
+    assert_raise RuntimeError, ~r/^Configuration Error/, fn ->
+      Conform.Translate.to_config(schema, [], conf)
+    end
   end
 end